made tabbing and properties subfolder, moved scripts
